evolution

+++ game-engine/src/main/kotlin/akostenko/aicars/GameSettings.kt
@@ -19,6 +19,8 @@ import org.lwjgl.input.Keyboard.KEY_Q
 import org.lwjgl.input.Keyboard.KEY_R
 import org.newdawn.slick.KeyListener
 import org.slf4j.LoggerFactory
+import java.io.BufferedInputStream
+import java.io.BufferedOutputStream
 import java.io.FileInputStream
 import java.io.FileOutputStream
 import java.io.IOException
@@ -29,12 +31,18 @@ import java.nio.file.Paths
 import java.nio.file.StandardOpenOption.CREATE_NEW
 import java.nio.file.StandardOpenOption.TRUNCATE_EXISTING
 import java.nio.file.StandardOpenOption.WRITE
+import java.nio.file.attribute.FileAttribute
+import java.nio.file.attribute.PosixFilePermission
 import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+import java.util.*
 import java.util.zip.ZipInputStream
 import java.util.zip.ZipOutputStream
 
 class GameSettings {
     private val log  = LoggerFactory.getLogger(this.javaClass)
+    private val timeFormat = DateTimeFormatter.ofPattern("yyyyMMdd-HHmmss")
+    private val linksSupported = System.getProperty("os.name")?.toLowerCase()?.contains("windows") == false
 
     var globalListeners: List<KeyListener> = emptyList()
         private set
@@ -50,23 +58,41 @@ class GameSettings {
 
     fun readPopulation(): List<NNDriver> {
         if (Files.exists(populationsPath) && Files.isDirectory(populationsPath)) {
-            val lastPopulation = Files.readSymbolicLink(lastPopulation)
-            if (Files.exists(lastPopulation)) {
-                val stream = ZipInputStream(FileInputStream(lastPopulation.toFile()))
-                return NeuralNet.deserializePopulation(IOUtils.readLines(stream, StandardCharsets.UTF_8))
+            val lastPopulation : Path? =
+            if (linksSupported) {
+                Files.readSymbolicLink(GameSettings.lastPopulation)
+            } else {
+                findLatestPopulation()
             }
+            if (lastPopulation != null && Files.exists(lastPopulation)) {
+                BufferedInputStream(FileInputStream(lastPopulation.toFile())).use {
+                    return NeuralNet.deserializePopulation(IOUtils.readLines(it, StandardCharsets.UTF_8))
+                }
+            }
+        } else {
+            Files.createDirectories(populationsPath);
         }
         log.warn("Population not found under $populationsPath. It will be generated.")
         return NeuralNet.generatePopulation()
     }
 
     fun savePopulation(population: List<NNDriver>) {
-        val path = populationsPath.resolve("p${LocalDateTime.now()}.zip")
-        val stream = ZipOutputStream(FileOutputStream(path.toFile()))
-        IOUtils.writeLines(NeuralNet.serializePopulation(population), System.lineSeparator(), stream, StandardCharsets.UTF_8)
+        val path = populationsPath.resolve("p${timeFormat.format(LocalDateTime.now())}.zip")
+        BufferedOutputStream(FileOutputStream(path.toFile())).use {
+            IOUtils.writeLines(NeuralNet.serializePopulation(population), System.lineSeparator(), it, StandardCharsets.UTF_8)
+        }
 
-        Files.deleteIfExists(lastPopulation)
-        Files.createSymbolicLink(lastPopulation, path)
+        if (linksSupported) {
+            Files.deleteIfExists(lastPopulation)
+            Files.createSymbolicLink(lastPopulation, path)
+        }
+    }
+
+    private fun findLatestPopulation(): Path? {
+        return Files.list(populationsPath)
+                .filter( { p -> p != null && p.startsWith("p") && p.endsWith(".zip")})
+                .sorted(reverseOrder())
+                .findFirst().orElse(null)
     }
 
     companion object {
